+++ src/i2c.h
@@ -19,7 +19,7 @@
 #include "em_letimer.h"
 #include "efr32bg13p632f512gm48.h"
 
-extern uint16_t read_data;
+extern uint32_t read_data;
 
 /*
  * Function to initialize i2c0
